RN-664: Moved utility functions from @tupaia/central-server SurveyType to @tupaia/database SurveyType
